Initial logging complete
Using the slog library, the initial changes for logging are complete. A global logger with a drain for console and one for a log file have been implement, and now all code within the worker and master modules can log information using the standard log crate macros.

Also fixed a bug in the Radio.Send method that I didn't detect before (since I was not logging), in which I was sending a Message structure rather than a MessageType(data) enum.

At this point, the processes comunicate fine (when they start up in the proper order) and respond to each other messages.