Further fixes for TCP endpoints.
The endpoints mostly work now for both modes, but it's a fucking mess, since they are created all over the place. This will need to be consolidated for better maintenance in the near future once the demo is done.

Radio.send is not being called in all instances for some reason, need to explore that.