Finished feature
As it turns out, there was no need to move the message-polling mechanism to another thread. After binding the socket to an address, any messages targerted for this node will be cached and can be retrieved later via reading from the socket.

Thus, starting the worker is now broken down in 3 parts:
1. Bind the listening address.
2. Do protocol initial messaging.
3. Loop listen for messages.

All implemented tests are passing,