Integration tests working
At this point, a basic integration test is added and works as expected.

However, usefulness is limited since the current way to evaluate success of the test is the exit code of the tested command and a 1 line evaluation from stdout.

There *should* be an easy way to write more complex tests by parsing the output, which should be easy to do since each line is properly formatted JSON.