Feature complete
The basic 2-radio support is in. This means that both radios are loaded from configuration (if they exist), passed to the protocol on creation and then start their respective listeners on their own thread. Protocols that use only 1 radio don't have to care about the 2nd radio support.

All tests are passing, but no new tests have been added to stress the inner-workings of a protocol that actually uses both radios. This will be added in an upcoming feature branch.