Feature complete
The protocol is now functioning and complete, and uses both radios.
The new experimental test  "tmembership_advanced_sustained" exercises this protocol.

The test remains experimental as it remains to be integrated with the Rust testing framework. Instructions on how to run the test and evaluate it are contained in tests/experiments/notes.md