Feature complete
All tests are passing (naive_basic is ignored at the moment to enable pikachu). The basic infrastructure for the commands is enabled and the Send command for naive_routing is already wired-in. The only thing left is to code the send command inside the protocol. Will handle in another feature.