Feature complete
The entire project has been moved to use the latest slog library (2.4.2 at the moment).
This change brings about several changes:

1. The project no longer relies on slog_scope to set global loggers.
2. Instead, loggers are now created by a dedicated function and passed around.
3. To keep the solution viable, the loggers are wrapped around an instance of
   slog_async, which allows me to safely share references to the logger across
   threads.
4. The use of slog_async has 2 more implications:
  4.1 First, it should improve overall performance of simulations as it pushes logging
      to a dedicated thread, no longer blocking execution threads.
  4.2 This also has consequences. When all of the threads scheduled by the master are
      finished and the test is ready to conclude, the logging thread might still be busy.

Since we have no direct hande to the logging thread used by slog_async, a preventive
measure has been put in place to check if the log continues to grow in size for up to
2 full seconds. This seems to suffice at the moment for all test logs.

The new features of the logging crate might enable me to refine how I log my data for
future analysis. This will be explored in subsequent transactions.