Feature complete
The following changes are now implemented:
* A new parameter has been introduced to worker_cli (register_worker).
* When register_worker is true, the worker handles its own registration to the database (same as before).
* When false, the master handles the registration.
* This is done when the master runs the corresponding worker process.

These changes prevent the significant contention that was created when multiple processes started at the same time (which is the case for all tests).

A bug was also fixed in mobility::stop_workers. This function only worked when the passed list of workers contained only 1 worker. This has been fixed.