Changes complete
All relevant code paths have been changed to use the new MeshSim error,
rather than the previous, module-dependant errors. The error handling
strategy now depends on using map_err instead of implementing the From
trait, in order to capture local error information (e.g. what was the
code doing when the error occurred) as well as appending the cause
when available.

All but one test are passing (route_teardown). Have elected to fix that
test in a future branch since it's failing with a consistent output, so
it's likely to not be a test error but a code one.