Feature complete
RGR-II now implements a heuristic to improve the efficiency of the route-discovery
of RGR-I.

An element that remains to be fixed is that the code that updates the vicinity cache
in the process_route_discovery_msg function is commented out. This is because with it,
the test fails (not all messages arrive at their destination). This is counter-intuitive,
as more data should promote more msg propagation. However, the opposite seems to hapen here.

Current theory is that either the broadcast is causing trouble for this (e.g. msg-delivery
order), which will be addressed in the next feature branch; or that having more message
propagation leads to nodes hearing about a route from different sources, thus, changing
the internal state of the random probability. Investigation is needed.

All tests are passing.