Fix complete
Delay in the read_message method of SimulatedListener is now
inverserly proportional to the order in which the message is
received as part of the broadcast set of messages.

A test added for debugging, but currently ignored since results
are not deterministic, so no good way to test it.

All tests passing.