Implementation of RERR
The implementation of the route-break detection is mostly done. However, it is not finished.
The main issue is that it the performance of the protocol falls significantly. More debugging
is needed, but committing the work now since the changeset is growing too large and becoming
more difficult to track the changes.