Logging modification
While I initially set out to make log_tx to take a generic parameter rather than a boxed trait object,
that proved too difficult as the protocols themselves are boxed objects.

Instead, a comprise was reached by creating a new enum of Protocol messages. This enum of enums holds
a variation for each of the protocols and has an associated type of that protocol's Message enum.

This allows me to pass around ProtocolMessage variables that ultimately hold the message to be logged.
slog::KV is now implemented in this new uber Enum.

This changes should impact performance possitively as no more boxing/unboxing or cache misses will happen
when logging packets, but this remains to be verified in the experiment servers.

All tests are passing.