Env_var DB connections
Modified the get_db_connection function to pick up the connection string from an
environment variable rather than from an .env file. This change follows an observation
in experiments in which rarely, but definitely, nodes would fail to connect to the DB
from what appears to be contention on the .env file. Having the connection string
set as an environment variable should address this.

All tests are passing.