NetworkContention errorkind
Added a new NetworkContention errorkind to MeshSim::ErrorKind. This variant is used to
signal when a broadcast operation has failed due to medium contention. Adding to that,
simulated radio no longer propagates such errors, just logs the event.

This pushed me to change the API of broadcast() to return an Option<TX> to signal a
transmission did not occur, but no error happened. This is not elegant but it is functional.
Might be changed later.

All tests are passing.