Removed VALID_ROUTE flag
Removed the aforementiond flag. the RFC refers to a VALID_ROUTE and a VALID_SEQ_NO flag
in different parts, but after much reviewing, they refer to the same concept.

Pushing this change to test in the simulation server as this is potentially a massive change.

All tests are passing.