Data analysis integrated into test specs
Since the processing of the experimental results depends on greping the logs for patterns that are different
for each protocol, they have now been integrated into the test specs. I was unsure about this, since this
is more related to the experimental design than to the tool, but since those patterns don't change unless the
source code changes, it made sense to integrated them with the protocol specifications.