Removed the async transmission from the worker.
It was a mistaken abstraction to add asynchronous transmission for response messages in the worker.
While I still keep the outgoing queue for transmission, it is now done by the listening thread.
To accommodate for the extra processing, the timeout from the read has been diminished from 100ms to 10ms,
and it is now expressed in microseconds in the config file in case I find it needs to be sub ms in production.

All tests are passing, and the success rate is actually much higher.