Replaced radio handle with outqueue
BIG change. Realised that by sharing the radio handle with the protocols, they could transmit at the
same time as the worker main thread, which is a waste since one of them would lock. Instead, now I
implemented a shared outbound_queue to which the protocols push messages.

This change simplified the code and made it cleaner.

All tests are passing.