New worker registration process and worker commands
This is a massive change that modifies two architectural decisions of the platform:

1. Modify the way workers are registered in the system: Instead of having them output their address to stdout and have the master read them, they now send a UDP packet to the new master registration server. While the previous version worked perfectly well, it was necessary to do given how interlated the entire code base was.
2. Modify the way the workers receive commands from upper layers. Previously, the workers would and parse strings from their stdin to interpret as commands. To this effect, the Master spawned the processes and kept a handle to their stdin. These read processes were ugly as hell. The new way involves the worker reading from a command unix socket. These reads happen periodically by the worker main thread. Workers no longer spawn a command thread.

All tests are passing and with much more consistency than before.

There is potential for a bug in the Master registration process due to the fact that it uses UDP sockets. If errors present themselves in teh future with missing nodes, this is the first place to look for the cause.