Registration server switched to TCP
After testing in production the previous change, it was far from hardened, even though all tests passed locally.
Looking into the error, it seems at the high level of parallelism that the experiments have, the registration server
design did not hold up, as many of the nodes went unregistered, and worst of all, silently.

This has been addressed by making sure the registration server receives the registration command from workers, or the workers
error out. While this is now guarantee of the correct registration, is much better than before. Further, several portions of
the registration process have been hardened to be more efficient and resilient to errors.

All tests are passing locally successfully, but a one more commit is still expected to finalise the details of this feature.
Pushing this change in the mean time to get some data from the simulation server.