Worker processes cleanup
Added a new section at the end of the run_test method to make sure any remaining worker
processes are killed. Any process that remains after if should is logged as an error,
as this might be an indication of a deadlock or some other bug.

All tests are passing.