AODV deadlock
Ok, finally fixed the fucking deadlock. After further debugging based on the logs, I found that
when handling a DATA packet that triggered a route_repair, while holding a lock to the route
table, I passed a reference to the route_repair function, which immediately tried to get that
same lock too. Trying to acquire a lock that I already hold, results in a deadlock.

Further, some fixes to radio.rs. Added the msg_id to the log entry when packets are dropped due
to contention. This helps keep count of where all data packets end up. Further, realised the
exponent for the sleep calculation was starting at 1, where it should have been starting at 0.