Further AODV fixes
Removed an issue in which an intermediate node can reply to an RREQ if it has a valid route, regardless of whether it’s active or not. This could cause packets to drop later.

Also made sure that the ACTIVE_ROUTE and VALID_SEQ_NO flags are used consistently.

Other minor fixes.